+++ client/components/modal/ProductInfo.jsx
