+++ client/components/nav/RatingSnapshot.jsx
