+++ client/components/review/ReviewList.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable no-underscore-dangle */
 import React from 'react';
 import ReviewListItem from './ReviewListItem';
