+++ client/components/util/FitBar.jsx
